Product data types edited

+++ NetCoreBackend.Entities/Concrete/Product.cs
@@ -8,7 +8,7 @@ namespace NetCoreBackend.Entities.Concrete
         public string ProductName { get; set; }
         public int CategoryID { get; set; }
         public string QuantityPerUnit { get; set; }
-        public int UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
+        public decimal UnitPrice { get; set; }
+        public short UnitsInStock { get; set; }
     }
 }
